Dismantled the mess
