--- Sources/iOS/Indicators/IndicatorsConfiguration.swift
@@ -1,61 +1,0 @@
-/*
- 
- Project: PageViewer
- File: IndicatorsConfiguration.swift
- Created by: Egor Boyko
- Date: 29.08.2023
- 
- Status: #Complete | #Not decorated
- 
- */
-
-import SwiftUI
-
-//public struct IndicatorsConfiguration<Style: IndicatorStyle>: View {
-//    let range: Range<Int>
-//    let index: Int?
-//    let style: () -> Style
-//    public var body: some View {
-//        ForEach(range, id: \.self){ index in
-//            IndicatorStorage(
-//                index: index,
-//                selected: index == self.index,
-//                style: style
-//            )
-//        }
-//    }
-//}
-
-
-//public protocol IndicatorsConfiguration {
-//    associatedtype Body: View
-//    func makeConfiguredPageView<PageView: View, Indicators: View>(
-//        content: PageView,
-//        indicators: Indicators) -> Body
-//}
-//
-//extension IndicatorsConfiguration {
-//
-//    static func configured<PageView: View, Indicators: IndicatorStyle>(range: Range<Int>, index: Int?, indicators: Indicators, pageView: PageView) -> some View {
-//
-//
-//
-//
-//    }
-//
-//}
-
-//public struct IndicatorsConfiguration<Style: IndicatorStyle>: View {
-//    let range: Range<Int>
-//    let index: Int?
-//    let style: () -> Style
-//    public var body: some View {
-//        ForEach(range, id: \.self){ index in
-//            IndicatorStorage(
-//                index: index,
-//                selected: index == self.index,
-//                style: style
-//            )
-//        }
-//    }
-//}
