+++ Sources/iOS/PageView/PageView+Modifiers.swift
@@ -32,49 +32,28 @@ extension PageView {
         view.looped = value
         return view
     }
-    
-    public func indicators<Style: IndicatorStyle>(_ style: Style) -> some View {
-        style.makeConfiguredPageView(content: { AnyView(self) }){
-            AnyView(
-                ForEach(0..<self.views.count, id: \.self){ index in
-                    style.makeIndicator(
-                        selected: index == self.index?.wrappedValue,
-                        index: index
+    ///Wraps the view and adds indicator pages according to the style.
+    ///- Parameter style: Defines the appearance of the indicators and their position.
+    ///- Parameter maxView: Specifies the maximum number of Views at which indicators will be shown,
+    /// if the limit is not required, specify nil.
+    ///- Note: Specify this modifier last. it returns some View
+    @ViewBuilder
+    public func indicators<Style: IndicatorStyle>(
+        _ style: Style = DefaultIndicatorStyle(),
+        maxView: Int? = 10) -> some View {
+            if let maxView, self.views.count > maxView {
+                self
+            } else {
+                style.makeConfiguredPageView(content: { AnyView(self) }){
+                    AnyView(
+                        ForEach(0..<self.views.count, id: \.self){ index in
+                            style.makeIndicator(
+                                selected: index == self.index?.wrappedValue,
+                                index: index
+                            )
+                        }
                     )
                 }
-            )
-        }
-    }
-    
-    
-    
-    //    public func indicators<Style: IndicatorStyle, Body: View>(
-    //        _ style: Style,
-    //        _ content: (
-    //            _ content: Self,
-    //            _ indicators: IndicatorsConfiguration<Style>
-    //        ) -> Body) -> some View {
-    //        content(
-    //            self,
-    //            IndicatorsConfiguration(
-    //                range: 0..<self.views.count,
-    //                index: self.index?.wrappedValue,
-    //                style: { style }
-    //            )
-    //        )
-    //    }
-}
-struct Configuration<Style: IndicatorStyle>: View {
-    let range: Range<Int>
-    let index: Int?
-    let style: () -> Style
-    public var body: some View {
-        ForEach(range, id: \.self){ index in
-            IndicatorStorage(
-                index: index,
-                selected: index == self.index,
-                style: style
-            )
-        }
+            }
     }
 }
